:ambulance: fixed indices for queries

+++ models/LinkPredictor.py
@@ -72,7 +72,8 @@ class LinkPredictor(L.LightningModule):
 
         loss = self.loss_fn(y_hat, y)
         y_hat = F.sigmoid(y_hat)
-        return RecSysStepOutput(y, y_hat, loss, torch.zeros(y.shape[0]).to(y))
+        return RecSysStepOutput(y, y_hat, loss,
+                                torch.zeros(y.shape[0]).to(y).to(torch.long))
 
     def training_step(self, batch: HeteroData, batch_idx: int) -> STEP_OUTPUT:
         y, y_hat, loss, index = self.common_step(batch)
